+++ gulpfile.js/tasks/comp-lib.js
@@ -6,15 +6,12 @@ var gulp = require('gulp')
 var exec = require('child_process').exec
 var config = require('../config')
 var compLibConfig = require('../../component-lib.json')
-var componentParser = require('../util/component-library/componentParser')
-var componentRenderer = require('../util/component-library/componentRenderer')
-var updateLibraryNav = require('../util/component-library/updateLibraryNav')
 
 gulp.task('clean-comp-lib', function (cb) {
   del(compLibConfig.destination, cb)
 })
 
-gulp.task('kss-node', ['clean-comp-lib', 'sass', 'images', 'browserify'], function (cb) {
+gulp.task('component-library', ['clean-comp-lib', 'sass', 'images', 'browserify'], function (cb) {
   var env = global.runmode
   var genCompLib = './node_modules/.bin/kss-node --config component-lib.json'
 
@@ -30,18 +27,4 @@ gulp.task('kss-node', ['clean-comp-lib', 'sass', 'images', 'browserify'], functi
 
     cb(err)
   })
-})
-
-gulp.task('component-library', ['kss-node'], function (cb) {
-  gulp.src(config.compLib.src)
-      .pipe(componentParser())
-      .pipe(componentRenderer({
-        template: path.join(compLibConfig.template, 'new-structure.html')
-      }))
-      .pipe(updateLibraryNav({
-        library: path.join(compLibConfig.destination)
-      }))
-      .pipe(gulp.dest(compLibConfig.destination))
-
-  cb()
 })